+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Replace key values as token
 - Linking and linked attributes in template graph
 - Default template-id and output-format in query
+- Running queries
 
 ### Changed
 - Move FHIR Packages to dktk-exporter